requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.1
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
